requirements.txt
flask
gunicorn
scikit-learn
numpy
pandas
PyPDF2
python-docx
python-dotenv